+++ Profiles/Feral_Wildstalker.lua
@@ -21,6 +21,7 @@ local Profile = {
     displayName = "Feral Wildstalker",
     specID = 103,
     markerSpell = 439531, -- Bloodseeker Vines (Wildstalker exclusive)
+    version = 1,
 
     state = {
         tigersFuryUntil = 0,
@@ -127,3 +128,10 @@ end
 ------------------------------------------------------------------------
 
 Engine:RegisterProfile(Profile)
+
+if TrueShot.CustomProfile then
+    TrueShot.CustomProfile.RegisterConditionSchema("Druid.Feral.Wildstalker", {
+        { id = "in_tigers_fury", label = "In Tiger's Fury", params = {} },
+        { id = "in_berserk",     label = "In Berserk",      params = {} },
+    })
+end